Using SNI errors as uint as far as possible (#3017)
(Also setting myself up for failure when merging LocalDBAPI, but whatever)

+++ src/Microsoft.Data.SqlClient/netfx/src/Microsoft/Data/SqlClient/LocalDBAPI.cs
@@ -90,7 +90,7 @@ namespace Microsoft.Data
                             {
                                 SniError sniError = new SniError();
                                 SNINativeMethodWrapper.SNIGetLastError(out sniError);
-                                throw CreateLocalDBException(errorMessage: StringsHelper.GetString("LocalDB_FailedGetDLLHandle"), sniError: (int)sniError.sniError);
+                                throw CreateLocalDBException(errorMessage: StringsHelper.GetString("LocalDB_FailedGetDLLHandle"), sniError: sniError.sniError);
                             }
                         }
                     }
@@ -224,13 +224,13 @@ namespace Microsoft.Data
         }
 
 
-        static SqlException CreateLocalDBException(string errorMessage, string instance = null, int localDbError = 0, int sniError = 0)
+        static SqlException CreateLocalDBException(string errorMessage, string instance = null, int localDbError = 0, uint sniError = 0)
         {
             Debug.Assert((localDbError == 0) || (sniError == 0), "LocalDB error and SNI error cannot be specified simultaneously");
             Debug.Assert(!string.IsNullOrEmpty(errorMessage), "Error message should not be null or empty");
             SqlErrorCollection collection = new SqlErrorCollection();
 
-            int errorCode = (localDbError == 0) ? sniError : localDbError;
+            int errorCode = (localDbError == 0) ? (int)sniError : localDbError;
 
             if (sniError != 0)
             {
